Bring calledWith messages more in line with Sinon's output (#1)
This changes the nonNegatedSuffix parameter to instead take a pair of suffixes, nonNegated and negated. This allows us to specify %D for the nonNegated case, which eliminates the noisy duplication, while still using %*%C for the negated case. This matches assert.js in core Sinon as best I can tell.